backend/pyproject.toml
[project]
name = "ai-legal-agent-backend"
version = "0.1.0"
description = "AI法务智能体系统后端服务"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "Apache-2.0" }
authors = [
    { name = "AI Legal Agent Team" }
]
keywords = ["ai", "legal", "multi-agent", "camel-ai", "fastapi"]

dependencies = [
    # Web框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    # 数据库
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    # 缓存与任务队列
    "redis>=5.0.1",
    "celery>=5.3.6",
    # 多智能体框架
    "camel-ai>=0.2.0",
    "mcp>=1.0.0",
    "jsonschema>=4.0.0",
    # 向量数据库
    "qdrant-client>=1.7.0",
    # 知识图谱
    "neo4j>=5.15.0",
    # 文档处理
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "python-pptx>=0.6.23",
    # AI/ML
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.7.0",
    "sentence-transformers>=2.3.1",
    # 工具
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    # 对象存储
    "minio>=7.2.3",
    # 数据处理
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "email-validator>=2.3.0",
    "playwright>=1.57.0",
    "beautifulsoup4>=4.14.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
